server/requirements.txt
python-socketio==5.11.2
aiohttp==3.9.5
aiokafka==0.10.0
